other/maps/requirements.txt
osmnx~=1.1.1
rasterio~=1.2.6
opencv-python~=4.5.2.54
GDAL~=3.2.2
matplotlib~=3.4.2
Pillow~=8.3.0
numpy~=1.21.0
tqdm~=4.61.1